refactor: Search bar extracted
